bug fixes, small design improvements, change kid id to int
- fix recording bug (when a word is typed before being recorded)
- fix item list screen so it changes action bar colors appropriately on
older phone versions
- move small profile pic in actionbar to vertical center
- change overflow icon to lighter color
- use int instead of long to store kid id everywhere, including shared
preferences

+++ src/com/nadajp/littletalkers/AddKidActivity.java
@@ -76,7 +76,7 @@ public class AddKidActivity extends Activity implements
       }
    }
 
-   public void onKidAdded(long kidId)
+   public void onKidAdded(int kidId)
    {
       Intent intent = new Intent(this, AddItemActivity.class);
       intent.putExtra(Prefs.CURRENT_KID_ID, kidId);
@@ -85,7 +85,7 @@ public class AddKidActivity extends Activity implements
       finish(); // unload this activity
    }
    
-   public void onKidUpdated(long kidId)
+   public void onKidUpdated(int kidId)
    {
       Intent intent = new Intent(this, ManageKidsActivity.class);
       startActivity(intent);
